+++ packages/brand-moment/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.2.3...@kaizen/brand-moment@1.2.4) (2021-06-25)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.2.2...@kaizen/brand-moment@1.2.3) (2021-06-24)
+
+
+### Bug Fixes
+
+* Add better RTL support to BrandMoment ([#1720](https://github.com/cultureamp/kaizen-design-system/issues/1720)) ([17426fc](https://github.com/cultureamp/kaizen-design-system/commit/17426fc521b089e1c640fad548e8e4b4dfe837e6))
+
+
+
+
+
+## [1.2.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.2.1...@kaizen/brand-moment@1.2.2) (2021-06-24)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.2.0...@kaizen/brand-moment@1.2.1) (2021-06-23)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
 # [1.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.1.3...@kaizen/brand-moment@1.2.0) (2021-06-22)