+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@2.9.0...@kaizen/design-tokens@2.9.1) (2021-06-25)
+
+**Note:** Version bump only for package @kaizen/design-tokens
+
+
+
+
+
 # [2.9.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@2.8.3...@kaizen/design-tokens@2.9.0) (2021-06-11)